requirements.txt
# Core Django backend
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-dateutil>=2.8.0

# AI / menu search / embeddings side
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pdfplumber>=0.9.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0

# Database (built-in to Python, listed just for clarity)
# sqlite3

# Optional: For better NLP
# spacy>=3.5.0
# nltk>=3.8.0

# Optional: For payment integration (if needed later)
# razorpay>=1.3.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# Core Django dependencies
django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0

# Chatbot & AI dependencies
groq>=0.4.0
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# PDF and image processing
pdf2image>=1.16.3
Pillow>=10.0.0

# Additional utilities
requests>=2.31.0
pydantic>=2.0.0

# Optional but recommended
gunicorn>=21.0.0  # For production deployment
redis>=5.0.0      # For caching (optional)
celery>=5.3.0     # For async tasks (optional)
